+++ src/lab.rs
@@ -4,17 +4,19 @@
 
 use std::cmp::{max, min};
 use std::panic::resume_unwind;
+use std::path::Path;
 use std::sync::Mutex;
 use std::thread;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use itertools::Itertools;
-use tracing::{debug, debug_span, error, info, trace, warn};
+use tracing::{debug, debug_span, error, trace, warn};
 
 use crate::cargo::run_cargo;
 use crate::outcome::LabOutcome;
 use crate::output::OutputDir;
 use crate::package::Package;
+use crate::timeouts::Timeouts;
 use crate::*;
 
 /// Run all possible mutation experiments.
@@ -32,7 +34,6 @@ pub fn test_mutants(
     console: &Console,
 ) -> Result<LabOutcome> {
     let start_time = Instant::now();
-
     if options.shuffle {
         fastrand::shuffle(&mut mutants);
     }
@@ -42,26 +43,23 @@ pub fn test_mutants(
         warn!("No mutants found under the active filters");
         return Ok(LabOutcome::default());
     }
-    let all_packages = mutants.iter().map(|m| m.package()).unique().collect_vec();
-    debug!(?all_packages);
+    let mutant_packages = mutants.iter().map(|m| m.package()).unique().collect_vec(); // hold
+    debug!(?mutant_packages);
 
     let output_mutex = Mutex::new(output_dir);
-    let build_dir = if options.in_place {
-        BuildDir::in_place(workspace_dir)?
-    } else {
-        BuildDir::copy_from(workspace_dir, options.gitignore, options.leak_dirs, console)?
+    let build_dir = match options.in_place {
+        true => BuildDir::in_place(workspace_dir)?,
+        false => BuildDir::copy_from(workspace_dir, options.gitignore, options.leak_dirs, console)?,
     };
-    let baseline_outcome = match options.baseline {
+
+    let timeouts = match options.baseline {
         BaselineStrategy::Run => {
             let outcome = test_scenario(
                 &build_dir,
                 &output_mutex,
                 &Scenario::Baseline,
-                &all_packages,
-                Timeouts {
-                    test: options.test_timeout.unwrap_or(Duration::MAX),
-                    build: options.build_timeout.unwrap_or(Duration::MAX),
-                },
+                &mutant_packages,
+                Timeouts::for_baseline(&options),
                 &options,
                 console,
             )?;
@@ -77,50 +75,43 @@ pub fn test_mutants(
                     .expect("lock output_dir")
                     .take_lab_outcome());
             }
-            Some(outcome)
+            Timeouts::from_baseline(&outcome, &options)
         }
-        BaselineStrategy::Skip => None,
+        BaselineStrategy::Skip => Timeouts::without_baseline(&options),
     };
-    let mut build_dirs = vec![build_dir];
+    debug!(?timeouts);
 
-    let baseline_duration_by_phase = |phase| {
-        baseline_outcome
-            .as_ref()
-            .and_then(|so| so.phase_result(phase))
-            .map(|pr| pr.duration)
-    };
-    let timeouts = Timeouts {
-        build: build_timeout(baseline_duration_by_phase(Phase::Build), &options),
-        test: test_timeout(baseline_duration_by_phase(Phase::Test), &options),
-    };
-
-    let jobs = max(1, min(options.jobs.unwrap_or(1), mutants.len()));
-    for i in 1..jobs {
-        debug!("copy build dir {i}");
-        build_dirs.push(BuildDir::copy_from(
-            workspace_dir,
-            options.gitignore,
-            options.leak_dirs,
-            console,
-        )?);
-    }
-    debug!(build_dirs = ?build_dirs);
-
+    let build_dir_0 = Mutex::new(Some(build_dir));
     // Create n threads, each dedicated to one build directory. Each of them tries to take a
     // scenario to test off the queue, and then exits when there are no more left.
     console.start_testing_mutants(mutants.len());
+    let n_threads = max(1, min(options.jobs.unwrap_or(1), mutants.len()));
     let pending = Mutex::new(mutants.into_iter());
     thread::scope(|scope| -> crate::Result<()> {
         let mut threads = Vec::new();
-        // TODO: Maybe, make the copies in parallel on each thread, rather than up front?
-        for build_dir in build_dirs {
+        for _i_thread in 0..n_threads {
             threads.push(scope.spawn(|| -> crate::Result<()> {
-                let build_dir = build_dir; // move it into this thread
-                trace!(thread_id = ?thread::current().id(), ?build_dir, "start thread");
+                trace!(thread_id = ?thread::current().id(), "start thread");
+                // First thread to start can use the initial build dir; others need to copy a new one
+                let build_dir_0 = build_dir_0.lock().expect("lock build dir 0").take(); // separate for lock
+                let build_dir = match build_dir_0 {
+                    Some(d) => d,
+                    None => {
+                        debug!("copy build dir");
+                        BuildDir::copy_from(
+                            workspace_dir,
+                            options.gitignore,
+                            options.leak_dirs,
+                            console,
+                        )?
+                    }
+                };
+                let _thread_span =
+                    debug_span!("worker thread", build_dir = ?build_dir.path()).entered();
                 loop {
                     // Extract the mutant in a separate statement so that we don't hold the
                     // lock while testing it.
-                    let next = pending.lock().map(|mut s| s.next());
+                    let next = pending.lock().map(|mut s| s.next()); // separate for lock
                     match next {
                         Err(err) => {
                             // PoisonError is not Send so we can't pass it directly.
@@ -129,7 +120,7 @@ pub fn test_mutants(
                         Ok(Some(mutant)) => {
                             let _span =
                                 debug_span!("mutant", name = mutant.name(false, false)).entered();
-                            let package = mutant.package().clone();
+                            let package = mutant.package().clone(); // hold
                             test_scenario(
                                 &build_dir,
                                 &output_mutex,
@@ -141,8 +132,7 @@ pub fn test_mutants(
                             )?;
                         }
                         Ok(None) => {
-                            trace!("no more work for this thread");
-                            return Ok(());
+                            return Ok(()); // no more work for this thread
                         }
                     }
                 }
@@ -191,78 +181,6 @@ pub fn test_mutants(
     Ok(lab_outcome)
 }
 
-#[derive(Copy, Clone)]
-struct Timeouts {
-    build: Duration,
-    test: Duration,
-}
-
-fn phase_timeout(
-    phase: Phase,
-    explicit_timeout: Option<Duration>,
-    baseline_duration: Option<Duration>,
-    minimum: Duration,
-    multiplier: f64,
-    options: &Options,
-) -> Duration {
-    const FALLBACK_TIMEOUT_SECS: u64 = 300;
-    fn warn_fallback_timeout(phase_name: &str, option: &str) {
-        warn!("An explicit {phase_name} timeout is recommended when using {option}; using {FALLBACK_TIMEOUT_SECS} seconds by default");
-    }
-
-    if let Some(timeout) = explicit_timeout {
-        return timeout;
-    }
-
-    match baseline_duration {
-        Some(_) if options.in_place && phase != Phase::Test => {
-            warn_fallback_timeout(phase.name(), "--in-place");
-            Duration::from_secs(FALLBACK_TIMEOUT_SECS)
-        }
-        Some(baseline_duration) => {
-            let timeout = max(
-                minimum,
-                Duration::from_secs((baseline_duration.as_secs_f64() * multiplier).ceil() as u64),
-            );
-
-            if options.show_times {
-                info!(
-                    "Auto-set {} timeout to {}",
-                    phase.name(),
-                    humantime::format_duration(timeout)
-                );
-            }
-            timeout
-        }
-        None => {
-            warn_fallback_timeout(phase.name(), "--baseline=skip");
-            Duration::from_secs(FALLBACK_TIMEOUT_SECS)
-        }
-    }
-}
-
-fn test_timeout(baseline_duration: Option<Duration>, options: &Options) -> Duration {
-    phase_timeout(
-        Phase::Test,
-        options.test_timeout,
-        baseline_duration,
-        options.minimum_test_timeout,
-        options.test_timeout_multiplier.unwrap_or(5.0),
-        options,
-    )
-}
-
-fn build_timeout(baseline_duration: Option<Duration>, options: &Options) -> Duration {
-    phase_timeout(
-        Phase::Build,
-        options.build_timeout,
-        baseline_duration,
-        Duration::from_secs(20),
-        options.build_timeout_multiplier.unwrap_or(2.0),
-        options,
-    )
-}
-
 /// Test various phases of one scenario in a build dir.
 ///
 /// The [BuildDir] is passed as mutable because it's for the exclusive use of this function for the
@@ -281,6 +199,11 @@ fn test_scenario(
         .expect("lock output_dir to create log")
         .create_log(scenario)?;
     log_file.message(&scenario.to_string());
+    let phases: &[Phase] = if options.check_only {
+        &[Phase::Check]
+    } else {
+        &[Phase::Build, Phase::Test]
+    };
     let applied = scenario
         .mutant()
         .map(|mutant| {
@@ -290,16 +213,12 @@ fn test_scenario(
             mutant.apply(build_dir)
         })
         .transpose()?;
-    console.scenario_started(scenario, log_file.path())?;
+    let dir: &Path = build_dir.path().as_ref();
+    console.scenario_started(dir, scenario, log_file.path())?;
 
     let mut outcome = ScenarioOutcome::new(&log_file, scenario.clone());
-    let phases: &[Phase] = if options.check_only {
-        &[Phase::Check]
-    } else {
-        &[Phase::Build, Phase::Test]
-    };
     for &phase in phases {
-        console.scenario_phase_started(scenario, phase);
+        console.scenario_phase_started(dir, phase);
         let timeout = match phase {
             Phase::Test => timeouts.test,
             Phase::Build | Phase::Check => timeouts.build,
@@ -315,7 +234,7 @@ fn test_scenario(
         )?;
         let success = phase_result.is_success(); // so we can move it away
         outcome.add_phase_result(phase_result);
-        console.scenario_phase_finished(scenario, phase);
+        console.scenario_phase_finished(dir, phase);
         if !success {
             break;
         }
@@ -326,157 +245,7 @@ fn test_scenario(
         .expect("lock output dir to add outcome")
         .add_scenario_outcome(&outcome)?;
     debug!(outcome = ?outcome.summary());
-    console.scenario_finished(scenario, &outcome, options);
+    console.scenario_finished(dir, scenario, &outcome, options);
 
     Ok(outcome)
-}
-
-#[cfg(test)]
-mod test {
-    use std::str::FromStr;
-
-    use indoc::indoc;
-
-    use super::*;
-    use crate::config::Config;
-
-    #[test]
-    fn timeout_multiplier_from_option() {
-        let args = Args::parse_from(["mutants", "--timeout-multiplier", "1.5"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.test_timeout_multiplier, Some(1.5));
-        assert_eq!(
-            test_timeout(Some(Duration::from_secs(40)), &options),
-            Duration::from_secs(60),
-        );
-    }
-
-    #[test]
-    fn test_timeout_unaffected_by_in_place_build() {
-        let args = Args::parse_from(["mutants", "--timeout-multiplier", "1.5", "--in-place"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(
-            test_timeout(Some(Duration::from_secs(40)), &options),
-            Duration::from_secs(60),
-        );
-    }
-
-    #[test]
-    fn build_timeout_multiplier_from_option() {
-        let args = Args::parse_from(["mutants", "--build-timeout-multiplier", "1.5"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.build_timeout_multiplier, Some(1.5));
-        assert_eq!(
-            build_timeout(Some(Duration::from_secs(40)), &options),
-            Duration::from_secs(60),
-        );
-    }
-
-    #[test]
-    fn build_timeout_is_affected_by_in_place_build() {
-        let args = Args::parse_from(["mutants", "--build-timeout-multiplier", "1.5", "--in-place"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(
-            build_timeout(Some(Duration::from_secs(40)), &options),
-            Duration::from_secs(300),
-        );
-    }
-
-    #[test]
-    fn timeout_multiplier_from_config() {
-        let args = Args::parse_from(["mutants"]);
-        let config = Config::from_str(indoc! {r#"
-            timeout_multiplier = 2.0
-        "#})
-        .unwrap();
-        let options = Options::new(&args, &config).unwrap();
-
-        assert_eq!(options.test_timeout_multiplier, Some(2.0));
-        assert_eq!(
-            test_timeout(Some(Duration::from_secs(42)), &options),
-            Duration::from_secs(42 * 2),
-        );
-    }
-
-    #[test]
-    fn build_timeout_multiplier_from_config() {
-        let args = Args::parse_from(["mutants"]);
-        let config = Config::from_str(indoc! {r#"
-            build_timeout_multiplier = 2.0
-        "#})
-        .unwrap();
-        let options = Options::new(&args, &config).unwrap();
-
-        assert_eq!(options.build_timeout_multiplier, Some(2.0));
-        assert_eq!(
-            build_timeout(Some(Duration::from_secs(42)), &options),
-            Duration::from_secs(42 * 2),
-        );
-    }
-
-    #[test]
-    fn timeout_multiplier_default() {
-        let args = Args::parse_from(["mutants"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.test_timeout_multiplier, None);
-        assert_eq!(
-            test_timeout(Some(Duration::from_secs(42)), &options),
-            Duration::from_secs(42 * 5),
-        );
-    }
-
-    #[test]
-    fn build_timeout_multiplier_default() {
-        let args = Args::parse_from(["mutants"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.build_timeout_multiplier, None);
-        assert_eq!(
-            build_timeout(Some(Duration::from_secs(42)), &options),
-            Duration::from_secs(42 * 2),
-        );
-    }
-
-    #[test]
-    fn timeout_from_option() {
-        let args = Args::parse_from(["mutants", "--timeout=8"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.test_timeout, Some(Duration::from_secs(8)));
-    }
-
-    #[test]
-    fn build_timeout_from_option() {
-        let args = Args::parse_from(["mutants", "--build-timeout=4"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.build_timeout, Some(Duration::from_secs(4)));
-    }
-
-    #[test]
-    fn timeout_multiplier_default_with_baseline_skip() {
-        // The --baseline option is not used to set the timeout but it's
-        // indicative of the realistic situation.
-        let args = Args::parse_from(["mutants", "--baseline", "skip"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.test_timeout_multiplier, None);
-        assert_eq!(test_timeout(None, &options), Duration::from_secs(300),);
-    }
-
-    #[test]
-    fn build_timeout_multiplier_default_with_baseline_skip() {
-        // The --baseline option is not used to set the timeout but it's
-        // indicative of the realistic situation.
-        let args = Args::parse_from(["mutants", "--baseline", "skip"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
-
-        assert_eq!(options.build_timeout_multiplier, None);
-        assert_eq!(build_timeout(None, &options), Duration::from_secs(300),);
-    }
 }